STL for chassis prototype 1
The main part of the chassis, will have a top part to cover the batteries, and will provide a place for the arduino and other components to sit.

This chassis uses a slide-on feature for the rear & nose parts of the car, they would just need the right sized tabs, they would slide in place before the top plate.

Design definitely has room for improvement.  Don't know if i like the slide-in tab design. Would like to not have to take of the top plate every time we need to take the nose or rear off( if we really need to).  Open to any ideas on how to improve this, or ideas for a different method.