Chassis top plate
added the top part that will cover the battery, and hold in the tabs.

Other thought was to make the top plate not block in the tabs, and use screws from the side of the chassis. LMK what you think could be better.

Arduino would sit on top of this, currently working on getting the arduino to be able to be either clipped in or screwed in place, alongside the breadboard.